% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces24.R
\docType{data}
\name{ces24}
\alias{ces24}
\title{Cooperative Election Study (CES) 2024 Dataset}
\format{
A data frame with 6000 rows and 694 variables:
\describe{
\item{case_id}{Unique respondent ID}
...
}
}
\source{
Cooperative Election Study, 2024
}
\usage{
ces24
}
\description{
This dataset contains responses from the 2024 CES survey,
with 60000 unweighted respondents and 694 variables.
}
\keyword{datasets}
